Hex function now works with long numbers.

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ccamargo <ccamargo@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/30 17:03:01 by ccamargo          #+#    #+#             */
-/*   Updated: 2022/06/10 14:39:05 by ccamargo         ###   ########.fr       */
+/*   Updated: 2022/06/10 20:12:49 by ccamargo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,10 @@
 # define FT_PRINTF_H
 
 # include <stdarg.h>
+# include "libft.h"
 
 int	ft_printf(const char *str, ...);
 int	ft_putunbr_fd(unsigned int n, int fd);
-int	ft_putnbr_hex_fd(int nbr, char *base, int fd);
+size_t	ft_putnbr_hex_fd(long nbr, char *base, int fd);
 
 #endif
